Fix/feat : score is now updated using the words per minute formula, csv document format has been modified

+++ Source/KeycubeDefender/Public/KCD_TypingStats.h
@@ -25,7 +25,7 @@ public:
 	}
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	double Score;
+	float Score;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
 	double TimeTaken;
